v43

+++ app/src/main/java/com/android/ground/ground/model/login/FacebookLoginUser.java
@@ -5,7 +5,7 @@ package com.android.ground.ground.model.login;
  */
 public class FacebookLoginUser {
     public int member_id;
-    public double id;
+    public long id;
     public int loginCode;
     public int clubYN;
     public int club_id;;
